requirements.txt
# Data fetching
yfinance==0.2.32
alpha-vantage==2.3.1
requests>=2.32.5
websocket-client==1.6.4

# Data processing - Updated for compatibility with existing packages
pandas==2.1.3
numpy>=1.26.0,<2.0.0  # Compatible with faiss-cpu, langchain, and trading system
pytz==2023.3

# Storage
psycopg2-binary==2.9.9
SQLAlchemy==2.0.23

# Streaming (Redis instead of Kafka)
redis==5.0.1

# Configuration
python-dotenv==1.0.0
pyyaml==6.0.1

# Utilities
schedule==1.2.0
loguru==0.7.2

# Testing
pytest==7.4.3
